requirements.txt
motor
loguru
pymongo
discord.py
dnspython
statuspageio